Move _WindowsAdapter to _utils module
